アイテムID/DBを自機からキャラクターにファイルを移動
