+++ CHAPTER6/キャラクタ/敵機/DB.py
@@ -17,22 +17,21 @@ class 機体ID(IntEnum):
     # アイテム
     アイテム    = auto()
 
-class DB:
-    #[0] 得点
-    #[1] 移動
-    #   [0] 発速度Y      ： 0～
-    #   [1] 速度X        ： 0～(0:横移動なし)
-    #   [2] 方向反転間隔 ： 左右反転の間隔をfpsで指定(0:反転しない)
-    #[2] 発射
-    #   [0] 間隔： fpsで指定(0:発射しない)
-    #   [1] 弾数： 同時に発射する弾数(0:発射しない/1/2/4)
-    #   [2] 速度： 移動速度(0:発射しない)
-    #[3] 破壊力
-    #   [0] 弾  ： 弾の攻撃で与えるダメージ
-    #   [1] 衝突： 体当たりで与えるダメージ
-    敵機 = {
-            機体ID.戦闘機1 : ( 30, ( 1.2, 0.0,  0 ), ( 40, 2, 1.6 ), ( 5 ,3 ) ),
-            機体ID.戦闘機2 : ( 90, ( 1.0, 1.2, 30 ), ( 40, 1, 2.0 ), ( 4 ,2 ) ),
-            機体ID.戦闘機3 : ( 60, ( 0.8, 0.0,  0 ), ( 80, 4, 1.3 ), ( 5 ,5 ) ),
-            機体ID.機雷    : (  0, ( 0.3, 0.8, 20 ), (  0, 0, 0.0 ), ( 0 ,0 ) ),
-            機体ID.アイテム: (  0, ( 0.3, 0.8, 20 ), (  0, 0, 0.0 ), ( 0 ,0 ) ) }
+#[0] 得点
+#[1] 移動
+#   [0] 発速度Y      ： 0～
+#   [1] 速度X        ： 0～(0:横移動なし)
+#   [2] 方向反転間隔 ： 左右反転の間隔をfpsで指定(0:反転しない)
+#[2] 発射
+#   [0] 間隔： fpsで指定(0:発射しない)
+#   [1] 弾数： 同時に発射する弾数(0:発射しない/1/2/4)
+#   [2] 速度： 移動速度(0:発射しない)
+#[3] 破壊力
+#   [0] 弾  ： 弾の攻撃で与えるダメージ
+#   [1] 衝突： 体当たりで与えるダメージ
+敵機DB = {
+        機体ID.戦闘機1 : ( 30, ( 1.2, 0.0,  0 ), ( 40, 2, 1.6 ), ( 5 ,3 ) ),
+        機体ID.戦闘機2 : ( 90, ( 1.0, 1.2, 30 ), ( 40, 1, 2.0 ), ( 4 ,2 ) ),
+        機体ID.戦闘機3 : ( 60, ( 0.8, 0.0,  0 ), ( 80, 4, 1.3 ), ( 5 ,5 ) ),
+        機体ID.機雷    : (  0, ( 0.3, 0.8, 20 ), (  0, 0, 0.0 ), ( 0 ,0 ) ),
+        機体ID.アイテム: (  0, ( 0.3, 0.8, 20 ), (  0, 0, 0.0 ), ( 0 ,0 ) ) }
