+++ CHAPTER6/キャラクタ/敵機/FN発射.py
@@ -3,7 +3,7 @@
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 import  pyxel
 import  処理.DB
-from    .DB     import 機体ID, DB
+from    .DB     import 機体ID, 敵機DB
 from    ..弾    import 弾発射
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
@@ -17,10 +17,10 @@ class 仕様クラス:
         else                        : キー = (引数_仕様.種類ID)
 
         self.衝突範囲   = (0, 0, 7, 7)
-        self.間隔       = DB.敵機[ キー ][2][0]
-        self.弾数       = DB.敵機[ キー ][2][1]
-        self.速度       = DB.敵機[ キー ][2][2]
-        self.破壊力     = DB.敵機[ キー ][3][0]
+        self.間隔       = 敵機DB[ キー ][2][0]
+        self.弾数       = 敵機DB[ キー ][2][1]
+        self.速度       = 敵機DB[ キー ][2][2]
+        self.破壊力     = 敵機DB[ キー ][3][0]
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃メイン
