+++ CHAPTER6/キャラクタ/敵機/FN移動.py
@@ -2,7 +2,7 @@
 #┃キャラクタ：敵機：移動機能
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 import  pyxel
-from    .DB import 機体ID, DB
+from    .DB import 機体ID, 敵機DB
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃仕様
@@ -14,9 +14,9 @@ class 仕様クラス:
         if 引数_仕様.アイテム区分   : キー = (機体ID.アイテム)
         else                        : キー = (引数_仕様.種類ID)
 
-        self.速度Y      = DB.敵機[ キー ][1][0]
-        self.速度X      = DB.敵機[ キー ][1][1]
-        self.反転間隔   = DB.敵機[ キー ][1][2]
+        self.速度Y      = 敵機DB[ キー ][1][0]
+        self.速度X      = 敵機DB[ キー ][1][1]
+        self.反転間隔   = 敵機DB[ キー ][1][2]
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃情報
