+++ CHAPTER6/キャラクタ/敵機/FN衝突.py
@@ -2,7 +2,7 @@
 #┃キャラクタ：敵機：衝突機能
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 import  処理.DB
-from    .DB     import 機体ID, DB
+from    .DB     import 機体ID, 敵機DB
 from    ..爆発  import 爆発開始
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
@@ -22,8 +22,8 @@ class 仕様クラス:
         else:
             self.衝突範囲   = (1, 1, 6, 6)
 
-        self.点数   = DB.敵機[ キー ][0]
-        self.破壊力 = DB.敵機[ キー ][3][1]
+        self.点数   = 敵機DB[ キー ][0]
+        self.破壊力 = 敵機DB[ キー ][3][1]
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃メイン
