--- CHAPTER6/キャラクタ/自機/DB.py
@@ -1,66 +1,0 @@
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃キャラクター：標的
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-from    enum import IntEnum, auto
-
-class アイテムID(IntEnum):
-
-    # 補給系
-    弾薬回復	    = 0
-    シールド回復	= auto()
-
-    # 永続効果
-    速度UP	        = auto()    # 0.5ずつ速くなる（上限あり）
-    発射_連射   	= auto()    # 永続効果
-    発射_貫通弾	    = auto()    # 永続効果
-    発射_弾数UP   	= auto()    # 永続効果、1ずつ増える（上限あり）
-
-    # 多方向同時発射
-    発射_上下方向   = auto()    # 発射_弾数UP と併用可
-    発射_左右方向   = auto()    # 発射_弾数UP と併用可
-    発射_4方向   	= auto()    # 発射_弾数UP と併用可
-    発射_8方向   	= auto()    # 発射_弾数UP は一時無効
-
-    # 敵からのダメージを緩和
-    ダメージ半減	= auto()    # ダメージが半減する
-    ダメージ無し	= auto()    # ノーダメージになる
-    ダメージ吸収    = auto()    # 敵の攻撃力をシールドに加える
-
-class 効果ID(IntEnum):
-
-    追加速度     = 1
-    連射         = auto()
-    防御UP       = auto()
-    貫通弾       = auto()
-    発射数       = auto()
-    発射方向     = auto()
-
-class DB:
-    #[0] 出現情報 
-    #   [0] 難易度     ： 難易度がこの数値異常で出現する 
-    #   [1] 出現率     ： 難易度がこの数値異常で出現する 
-    #[1] 詳細情報 
-    #   [0] 持続性     ： 効果が続く時間
-    #                   ・-1  ：使い切り(補給物資に使用する)
-    #                   ・0   ：永続
-    #                   ・1～ ：有限(秒)
-    #   [1] 登録値 
-    #       [0] 値      ： 登録する値 
-    #       [1] 上限値  ： 登録する値の上限 
-    アイテム = {
-        アイテムID.弾薬回復	        : ( ( 1,   0) , (-1, True       ) ),
-        アイテムID.シールド回復     : ( ( 2,   0) , (-1, True       ) ),
-
-        アイテムID.速度UP	        : ( ( 1,  80) , ( 0, ( 0.5 , 3 )) ),
-        アイテムID.発射_連射   	    : ( ( 2,  80) , ( 0, True       ) ),
-        アイテムID.発射_弾数UP      : ( ( 3,  80) , ( 0, ( 1 , 5 )  ) ),
-        アイテムID.発射_貫通弾	    : ( ( 4,  50) , (10, True       ) ),
-
-        アイテムID.発射_上下方向    : ( ( 5,  50) , (10, 1          ) ),
-        アイテムID.発射_左右方向    : ( ( 6,  30) , (10, 2          ) ),
-        アイテムID.発射_4方向   	: ( ( 7,  20) , (10, 3          ) ),
-        アイテムID.発射_8方向   	: ( ( 8,  10) , (10, 4          ) ),
-
-        アイテムID.ダメージ半減	    : ( ( 5,  20) , (10, 0.5        ) ),
-        アイテムID.ダメージ無し	    : ( ( 7,  10) , (10, 0          ) ),
-        アイテムID.ダメージ吸収     : ( (10,   5) , (10, -1         ) ) }
