+++ CHAPTER6/キャラクタ/自機/FN特殊.py
@@ -3,7 +3,7 @@
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 import  pyxel
 import  処理.DB
-from    .DB    import アイテムID, 効果ID, DB
+from    ..DB    import アイテムID, 効果ID, アイテムDB
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃情報
@@ -60,7 +60,7 @@ class 特殊クラス():
 		#┬
         #○アイテム種類IDと持続時間を退避する
         アイテムID  = 引数_アイテム.仕様.種類ID
-        持続時間    = DB.アイテム[ アイテムID ][1][0]
+        持続時間    = アイテムDB[ アイテムID ][1][0]
             #┴
         #│
         #◇┐アイテム区分を照会し効果を発動する
@@ -174,8 +174,8 @@ class 特殊クラス():
         #│ ▼処理を中断する
         #│
         #○基本情報を用意する
-        持続時間    = DB.アイテム[ 種類ID ][1][0]
-        固有値      = DB.アイテム[ 種類ID ][1][1]
+        持続時間    = アイテムDB[ 種類ID ][1][0]
+        固有値      = アイテムDB[ 種類ID ][1][1]
         #│
         #●発動済の同類効果を削除する
         self.Fn発動済を削除( 引数_効果ID )
@@ -198,9 +198,9 @@ class 特殊クラス():
         #│ ▼処理を中断する
         #│
         #○基本情報を用意する
-        持続時間    = DB.アイテム[ 種類ID ][1][0]
-        追加値      = DB.アイテム[ 種類ID ][1][1][0]
-        上限値      = DB.アイテム[ 種類ID ][1][1][1]
+        持続時間    = アイテムDB[ 種類ID ][1][0]
+        追加値      = アイテムDB[ 種類ID ][1][1][0]
+        上限値      = アイテムDB[ 種類ID ][1][1][1]
         #│
         #●同類(連射)の発動済を削除する
         既存値 = self.Fn発動済を削除( 引数_効果ID )
