+++ CHAPTER6/キャラクタ/自機/FN移動.py
@@ -2,8 +2,8 @@
 #┃キャラクタ：自機：移動機能
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 import  pyxel
-from    共通    import class入力操作    as 入力 
-from    .DB     import 効果ID
+from    共通    import class入力操作 as 入力 
+from    ..DB    import 効果ID
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃仕様
