+++ CHAPTER6/キャラクタ/自機/FN衝突.py
@@ -2,7 +2,7 @@
 #┃キャラクタ：自機：衝突機能
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 from 処理   import *
-from .DB    import 効果ID
+from ..DB   import 効果ID
 from ..爆発 import 爆発開始
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
