+++ CHAPTER6/処理/DB.py
@@ -3,6 +3,14 @@
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃ゲームに関する情報 
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+シーン      = None  # タイトル→プレイ→終了→タイトル...と移る
+プレイ時間  = 0     # 難易度の変化，出現タイミングの物差し
+難易度      = 0     # 敵の強さ、出現頻度の物差し
+得点        = 0     # プレイヤーの得点
+
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃インスタンス(キャラクタ用)
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 obj自機     = None
@@ -18,14 +26,6 @@ obj背景     = None
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 FN更新      = ()
 FNシーン    = {}
-
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃ゲームに関する情報 
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-シーン      = None  # タイトル→プレイ→終了→タイトル...と移る
-プレイ時間  = 0     # 難易度の変化，出現タイミングの物差し
-難易度      = 0     # 敵の強さ、出現頻度の物差し
-得点        = 0     # プレイヤーの得点
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃所有者に関するＩＤ
